backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
